+++ core/resources/src/correction.rs
@@ -6,9 +6,9 @@ use bevy::{
 #[derive(Debug, Event, Clone, Resource, Default)]
 pub struct StartCorrection {
     /// The tick at which correction simulation should sync. This tick does not get returned and re-applied in correction results.
-    pub start_tick: u64,
-    /// The last tick that is calculated. This should always be equal to TickRateStamp.large of the loop it gets called in.
-    pub last_tick: u64,
+    pub start_tick: u32,
+    /// The last tick that is calculated. This should always be equal to TickRatestamp.tick of the loop it gets called in.
+    pub last_tick: u32,
 }
 #[derive(Debug, Hash, PartialEq, Eq, Clone, SystemSet)]
 pub enum CorrectionServerSet {
@@ -16,7 +16,7 @@ pub enum CorrectionServerSet {
     SyncClear,
 }
 
-pub const MAX_CACHE_TICKS_AMNT: u64 = 32;
+pub const MAX_CACHE_TICKS_AMNT: u32 = 32;
 
 #[derive(Resource, Default)]
 pub struct IsCorrecting(pub bool);
